+++ src/executer.ts
@@ -1,12 +1,17 @@
-import { ExecaChildProcess, execaCommand } from 'execa';
-import { ExecuteMode, ExecuteOption, ResolvedCommonOptions } from './types/options';
-import { Logger } from '@farmfe/core';
+import { type ExecaChildProcess, execaCommand } from 'execa';
+import { ExecuteMode, type ExecuteOption, type ResolvedCommonOptions } from './types/options';
+import type { Logger } from '@farmfe/core';
 import { delay } from './util/async';
+import { trimEndLF } from './util/log';
 
 export class Executer {
     child?: ExecaChildProcess;
 
-    constructor(public option: ExecuteOption, public logger: Logger, public normalizedOption: ResolvedCommonOptions) {}
+    constructor(
+        public option: ExecuteOption,
+        public logger: Logger,
+        public normalizedOption: ResolvedCommonOptions,
+    ) {}
 
     execute(path: string, name: string, logger = this.logger) {
         switch (this.option.type) {
@@ -26,49 +31,34 @@ export class Executer {
     }
 
     async _execute(command: string, name: string, args: string[], logger: Logger) {
-        if (!this.child) {
+        if (this.child) {
             await this.closeChild();
         }
 
         const child = execaCommand([command, ...args].join(' '), {
             cwd: process.cwd(),
-            // TODO: proxy and use logger
             stdio: 'pipe',
         });
 
-        child.stdout?.on('data', (data) => {
-            logger.debug(data.toString());
-        });
+        child.stdout?.on('data', (data) => logger.debug(trimEndLF(data.toString())));
 
-        child.stderr?.on('data', (err) => {
-            logger.error(err);
-        });
+        child.stderr?.on('data', (err) => logger.error(err));
 
         this.child = child;
 
-        process.on('beforeExit', () => {
-            this.closeChild();
-        });
-        process.on('exit', () => {
-            this.closeChild();
-        });
+        process.on('beforeExit', this.closeChild);
+        process.on('exit', this.closeChild);
 
         child.on('exit', (code) => {
-            if (child) {
-                if (code === 0) {
-                    this.logger.info(`"${name}" PID ${child.pid} done`);
-                } else {
-                    this.logger.info(`"${name}" PID ${child.pid} killed`);
-                }
-                this.child = undefined;
-            }
+            this.logger.info(`"${name}" PID ${child.pid} ${code === 0 ? 'done' : `exit ${code}`}`);
+            this.child = undefined;
         });
     }
 
     async closeChild() {
         while (this.child && !this.child.killed) {
             this.child.kill();
-            await delay(300);
+            await delay(50);
         }
         this.child = undefined;
     }