+++ src/index.ts
@@ -1,5 +1,4 @@
 import {
-    logger,
     watch as farmWatch,
     start as farmStart,
     build as farmBuild,
@@ -7,14 +6,16 @@ import {
     getConfigFilePath,
     type FarmCLIOptions,
     type UserConfig,
+    version as FarmCoreVersion,
+    NoopLogger,
 } from '@farmfe/core';
 import cac from 'cac';
 import { readFileSync } from 'node:fs';
 import autoExecute, { NormalizeOption } from './plugins/auto-execute';
 import { ExecuteMode, type CommonOptions } from './types/options';
-import { autoExternal } from './plugins/auto-external';
+import autoExternal from './plugins/auto-external';
 import path from 'node:path';
-import { isString } from 'lodash-es';
+import { isBoolean, isString } from 'lodash-es';
 
 const { version } = JSON.parse(readFileSync(new URL('../package.json', import.meta.url)).toString());
 
@@ -40,36 +41,35 @@ function createInlineConfig(options: CommonOptions): InlineConfig {
     };
 }
 
-async function start(options: CommonOptions) {
-    const preNormalizeOption = await NormalizeOption.fromCommonOption(options, logger);
-
+async function autoStart(options: CommonOptions) {
+    const preNormalizeOption = await NormalizeOption.fromCommonOption(options, new NoopLogger());
     const inlineConfig = createInlineConfig(options);
-
     switch (preNormalizeOption.options.execute.type) {
         case ExecuteMode.Browser:
             return farmStart(inlineConfig);
         case ExecuteMode.Node:
         case ExecuteMode.Custom:
-            return farmBuild(inlineConfig);
+            return options.noWatch ? farmBuild(inlineConfig) : farmWatch(inlineConfig);
     }
+}
+
+async function start(options: CommonOptions) {
+    const inlineConfig = createInlineConfig(options);
+    await farmStart(inlineConfig);
 }
 
 async function watch(options: CommonOptions) {
     const inlineConfig = createInlineConfig(options);
-
-    await farmWatch(inlineConfig).then(() => {
-        logger.info('watch start...');
-    });
+    await farmWatch(inlineConfig);
 }
 
 async function build(options: CommonOptions) {
-    if (options.noExecute) {
+    if (options.noExecute && !options.noWatch) {
         await watch(options);
         return;
     }
 
     const inlineConfig = createInlineConfig(options);
-
     await farmBuild(inlineConfig);
 }
 
@@ -77,7 +77,7 @@ const cli = cac('farmup');
 
 cli.option(
     '--target [target]',
-    `target for output, default is node, support 'browser' | 'node' | 'node16' | 'node-legacy' | 'node-next' | 'browser-legacy' | 'browser-es2015' | 'browser-es2017' | 'browser-esnext'`
+    "target for output, default is node, support 'browser'、'node'、'node16'、'node-legacy'、'node-next'、'browser-legacy'、'browser-es2015'、'browser-es2017'、'browser-esnext'",
 )
     .option('--mode [mode]', 'mode for build, default is development, choose one from "development" or "production"')
     .option('--minify', 'minify for output')
@@ -85,44 +85,50 @@ cli.option(
     .option('--no-config', 'if farm.config.[ext] exists, it will be ignore')
     .option('--format [format]', 'choose one from "cjs" or "esm"')
     .option('--external [...external]', 'external')
-    .option('--watch [...files]', 'watch files')
-    .option('-w [...file]', 'watch files')
+    .option('-w, --watch [...files]', 'watch files', { default: false })
     .option('--no-auto-external', 'if not found module, auto as external', { default: true })
-    .option('--exec [file]', 'custom execute command')
-    .option('-e [file]', 'custom execute command');
+    .option('-e, --exec [file]', 'custom execute command');
 
 // biome-ignore lint/suspicious/noExplicitAny: <explanation>
 async function commonOptionsFromArgs(args: Record<string, any>): Promise<Partial<CommonOptions>> {
     const root = process.cwd();
-
     const configPath =
         typeof args.config === 'string'
             ? path.isAbsolute(args.config)
                 ? args.config
                 : path.resolve(root, args.config)
             : args.config
-            ? await getConfigFilePath(root)
-            : undefined;
-    const execute =
-        isString(args.exec) || isString(args.e) ? args.exec || args.e : args.exec === true ? undefined : undefined;
+              ? await getConfigFilePath(root)
+              : undefined;
+    const execute = isString(args.exec) && !isBoolean(args.exec) ? args.exec : undefined;
+
     return {
         root,
         target: args.target,
         args: args['--'],
         mode: args.mode,
         autoExternal: args.autoExternal,
-        execute: execute,
+        execute,
         format: args.format,
         config: configPath,
         minify: args.minify,
-
+        noWatch: args.watch === false,
         noExecute: args.exec === false,
-        watchFiles: [args.watch, args.w].flat().filter(Boolean),
+        watchFiles: [args.watch]
+            .flat()
+            .map((item) => (item === true ? undefined : item))
+            .filter(Boolean),
     };
 }
 
-cli.command('build [entry]', 'start watch for node or custom command')
+cli.command('[entry]', 'start ts/js/html file').action(async (entry, options) => {
+    autoStart({
+        entry: Array.isArray(entry) ? entry : [entry].filter(Boolean),
+        ...(await commonOptionsFromArgs(options)),
+    });
+});
 
+cli.command('build [entry]', 'start watch for node or custom command')
     .option('--no-exec', 'disable execute')
     .action(async (entry, options) => {
         build({
@@ -130,16 +136,17 @@ cli.command('build [entry]', 'start watch for node or custom command')
             ...(await commonOptionsFromArgs(options)),
         });
     });
-
-cli.command('[entry]', 'start watch for node or custom command').action(async (entry, options) => {
+cli.command('start [entry]', 'start server for html file').action(async (entry, options) => {
     start({
         entry: Array.isArray(entry) ? entry : [entry].filter(Boolean),
         ...(await commonOptionsFromArgs(options)),
     });
 });
 
-cli.help();
+cli.showHelpOnExit = true;
 
-cli.version(version);
+cli.showVersionOnExit = true;
+
+cli.version(`${version} (core: ${FarmCoreVersion})`);
 
 cli.parse();