+++ src/plugin-entry.ts
@@ -1,2 +1,4 @@
 export * from './plugins/auto-execute';
-export { default } from './plugins/auto-execute';
+export { default as autoExecute } from './plugins/auto-execute';
+export { default as autoExternal } from './plugins/auto-external';
+export { CLI_NAME, logger } from './config/constant';