+++ src/plugins/auto-external.ts
@@ -1,9 +1,9 @@
 import type { JsPlugin, PluginResolveHookResult } from '@farmfe/core';
 import { CLI_NAME, logger } from '../config/constant';
 
-export function autoExternal(): JsPlugin {
+export default function autoExternal(): JsPlugin {
     return {
-        priority: -Number.NEGATIVE_INFINITY,
+        priority: Number.NEGATIVE_INFINITY,
         name: `${CLI_NAME}:AutoExternal`,
         resolve: {
             filters: {