+++ src/util/file.ts
@@ -1,4 +1,4 @@
-import { stat } from "fs-extra";
+import { stat } from 'fs-extra';
 
 export const isExists = async (filename: string) => {
     try {