+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # farmup
 
+## 0.0.6
+
+### Patch Changes
+
+-   65d0577: 1. short option format 2. format debug info
+
+## 0.0.5
+
+### Patch Changes
+
+-   145c5e3: watch short option cannot enable watch mode
+
+## 0.0.4
+
+### Patch Changes
+
+-   804e90a: 1. fmt 2. child kill condition 3. add start command for html 4. default no watch mode
+
+## 0.0.3
+
+### Patch Changes
+
+-   9dcfa65: remove warn & auto external priority
+
 ## 0.0.2
 
 ### Patch Changes